revert

+++ src/js/ast.zig
@@ -1195,7 +1195,7 @@ pub const JSXEmptyExpression = struct {};
 pub const JSXText = struct {
     raw_start: u32,
     // len of the jsx text is u32 to prevent overflow, unlike other nodes len property which is u16 and enough
-    raw_len: u32,
+    raw_len: u16,
 };
 
 /// `{...children}`
